+++ kernel/arch/dreamcast/include/dc/video.h
@@ -297,7 +297,7 @@ void vid_flip(int32_t fb);
     
     \return                 Old border color value (RGB888)
 */
-uint32_t vid_border_color(int r, int g, int b);
+uint32_t vid_border_color(uint8_t r, uint8_t g, uint8_t b);
 
 /** \brief   Clear the framebuffer.
     \ingroup video_fb
@@ -314,7 +314,7 @@ uint32_t vid_border_color(int r, int g, int b);
     \param  g               The green value of the color (0-255).
     \param  b               The blue value of the color (0-255).
 */
-void vid_clear(int r, int g, int b);
+void vid_clear(uint8_t r, uint8_t g, uint8_t b);
 
 /** \brief   Clear VRAM.
     \ingroup video_vram
